Add files via upload
This is a fork of Gunar Schorcht's (gschorcht) code modified to use kernel-generated GPIO pin/line names (to make it easier to relate them to the gpiochip device) and to not auto-export gpio entries in the /sys/class/gpio directory. Also make sure that the i2c-dev module is loaded.